+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 ## Change log
 ----------------------
 
-Version 6.3-SNAPSHOT
+Version 6.4-SNAPSHOT
+-------------
+
+CHANGED:
+
+- update of parent version to 5.3
+- update of silly-collections version to 5.5.1
+- update of jobj-core version to 3.3
+- update of equalsverifier version to 3.1.10
+- bug in RandomPrimitivesExtensions.randomIntBetween fixed and unit tests extended 
+
+Version 6.3
 -------------
 
 ADDED: